transformers_smolagents_example/requirements.txt
smolagents[litellm]
python-dotenv
